+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `navigator.getInstalledRelatedApps()` shim
 - Add `UTM` generator and validator class, extending `URL`
 - `Promise.prototype.finally` shim
+- `preload()` function, creating a `<link rel="preload">`
 
 ## [v2.5.1] - 2020-08-21